Update rooms.json (#657)
The embed link wasn't updated in the master document for the Sacristy, so I cut the last part of the domain for the youtube link, into the current embed link. Assuming the youtube video in the master doc was updated, then this *should* be the correct embed information. 

I'll doublecheck if the doc is updated later and delete this PR if it's inaccurate.